requirements.txt
playwright==1.40.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
pyyaml==6.0.1